Split Cards into cards.{cpp,h}
